streamlit-app/requirements.txt
streamlit
psycopg2-binary
pandas
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
deprecation==2.1.0
exceptiongroup==1.2.2
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
gotrue==2.9.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.1.0
narwhals==1.9.0
numpy==2.1.1
packaging==24.1
pillow==10.4.0
postgrest==0.17.0
protobuf==5.28.2
pyarrow==17.0.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.2
realtime==2.0.5
referencing==0.35.1
requests==2.32.3
rich==13.9.1
rpds-py==0.20.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
storage3==0.8.0
streamlit==1.38.0
StrEnum==0.4.15
supabase==2.8.1
supafunc==0.6.0
tenacity==8.5.0
toml==0.10.2
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==4.0.2
websockets==13.1
yarl==1.13.1
matplotlib
dotenv